+++ graphtransliterator/transliterators/bundled.py
@@ -1,9 +1,4 @@
-# from functools import wraps
-# from graphtransliterator.core import # GraphTransliterator, GraphTransliteratorSchema
-# from graphtransliterator.graphs import VisitLoggingDirectedGraph
 from graphtransliterator.core import GraphTransliterator, CoverageTransliterator
-
-# from graphtransliterator.coverage import CoverageTransliterator
 import os
 import sys
 import yaml
@@ -11,10 +6,28 @@ import yaml
 
 class Bundled(CoverageTransliterator, GraphTransliterator):
     """
-    Subclass of GraphTransliterator used by a bundled GraphTransliterator.
+    Subclass of GraphTransliterator used for bundled Graph Transliterator.
     """
 
-    def _init_from(self, method=None, **kwargs):
+    @property
+    def directory(self):
+        """Directory of bundled transliterator, used to load settings."""
+        return self._module_dir()
+
+    @property
+    def name(self):
+        """Name of bundled transliterator, e.g. 'Example'"""
+        return self._module_name()
+
+    def _module_dir(self, **kwargs):
+        """Returns directory of module. Overwritten during testing."""
+        return os.path.dirname(sys.modules[self.__module__].__file__)
+
+    def _module_name(self):
+        """Returns name of module. Overwritten during testing."""
+        return self.__module__
+
+    def init_from(self, method=None, **kwargs):
         """Initialize from easy-reading YAML or from JSON."""
         # Save initialization data in case it becomes a CoverageTransliterator
         # self.yaml_test_file = os.path.join(
@@ -24,21 +37,15 @@ class Bundled(CoverageTransliterator, GraphTransliterator):
         # )
         # self.orig_module = self.__module__
         filename = os.path.join(
-            os.path.dirname(sys.modules[self.__module__].__file__),
-            self.__module__ + "." + method,  # error if None
+            self.directory, self.name + "." + method  # error if None
         )
         # Create GraphTransliterator using factory
         if method == "yaml":
             gt = GraphTransliterator.from_yaml_file(filename, **kwargs)
         elif method == "json":
             with open(filename, "r") as f:
-                gt = GraphTransliterator.loads(
-                    f.read(), **kwargs
-                )  # create GraphTransliterator object
-        else:
-            raise ValueError("Unknown method")
-        #
-        # # Select coverage superclass, if coverage set.
+                gt = GraphTransliterator.loads(f.read(), **kwargs)
+        # Select coverage superclass, if coverage set.
         if kwargs.get("coverage"):
             _super = CoverageTransliterator
         else:
@@ -65,44 +72,23 @@ class Bundled(CoverageTransliterator, GraphTransliterator):
         )
 
     def from_YAML(self, check_ambiguity=True, coverage=True, **kwargs):
-        """Initialize from bundled YAML file (best for development)."""
-        self._init_from(
+        """Initialize from bundled YAML file (best for development).
+
+        Parameters
+        ----------
+        check_ambiguity: `bool`
+            Should ambiguity be checked.
+        """
+        self.init_from(
             method="yaml", check_ambiguity=check_ambiguity, coverage=coverage, **kwargs
         )
 
     def from_JSON(self, check_ambiguity=False, coverage=False, **kwargs):
         """Initialize from bundled JSON file (best for speed)."""
-        self._init_from(
+        self.init_from(
             method="json", check_ambiguity=check_ambiguity, coverage=coverage, **kwargs
         )
 
-    #
-    # @visitlogging
-    # def init_from_bundled_yaml(self, extension=".yaml", **kwargs):
-    #     """Return a GraphTransliterator from a bundled easy-reading YAML file."""
-    #
-    #     yaml_filename = os.path.join(
-    #         os.path.dirname(sys.modules[self.__module__].__file__),
-    #         self.__module__ + extension
-    #     )
-    #     import pdb; pdb.set_trace()
-    #     gt = self.from_yaml_file(yaml_filename)
-    #     self.__init__(gt)
-    #     #return cls(gt)   #cls.from_yaml_file(yaml_filename))
-    #
-    # @classmethod
-    # @visitlogging
-    # def from_bundled_json(self, path, name, extension=".json", **kwargs):
-    #     """Return a GraphTransliterator from a bundled serialized JSON file."""
-    #     json_filename = os.path.join(
-    #         os.path.dirname(sys.modules[self.__module__].__file__),
-    #         self.__module__ + extension
-    #     )
-    #     with open(json_filename, "r") as f:
-    #         json_ = f.read()
-    #     return self.load(json_, **kwargs)
-
-    @classmethod
     def load_yaml_tests(self):
         """Iterator for YAML tests.
 
@@ -110,13 +96,10 @@ class Bundled(CoverageTransliterator, GraphTransliterator):
         `NAME_tests.yaml, e.g. `source_to_target/tests/source_to_target_tests.yaml`.
         """
         test_file = os.path.join(
-            os.path.dirname(sys.modules[self.__module__].__file__),
-            "tests",
-            self.__module__ + "_tests.yaml",
+            self.directory, "tests", "{}_tests.yaml".format(self.name)
         )
-        # test_file = self.yaml_test_file
         with open(test_file, "r") as f:
-            return yaml.safe_load(f)
+            return {str(k): str(i) for k, i in yaml.safe_load(f).items()}
 
     def run_tests(self, transliteration_tests):
         """Run transliteration tests.
@@ -127,6 +110,8 @@ class Bundled(CoverageTransliterator, GraphTransliterator):
             Dictionary of test from source -> correct target.
         """
         for source, target in transliteration_tests.items():
+            source = str(source)  # covert to str
+            target = str(target)
             result = self.transliterate(source)
             assert (
                 self.transliterate(source) == target