+++ graphtransliterator/transliterators/example/__init__.py
@@ -7,14 +7,14 @@ class Example(Bundled):
     """
 
     def __init__(self, **kwargs):
-        """Initializes transliterator from YAML or JSON (quicker)."""
+        """Initialize transliterator from YAML or JSON (quicker)."""
 
-        # While testing, initialize from YAML and check ambiguity.
+        # While testing, initialize from YAML and check ambiguity:
 
         self.from_YAML(
             **kwargs
         )  # defaults to check_ambiguity=True, check_coverage=True
 
-        # When set, cut the previous line and initialize more quickly from JSON
+        # When ready, cut the previous line and initialize more quickly from JSON:
 
-        # self.init_from_JSON() # check_ambiguity=False, check_coverage=False
+        # self.init_from_JSON(**kwargs) # check_ambiguity=False, check_coverage=False