+++ tests/test_transliterators.py
@@ -1,74 +1,50 @@
 # -*- coding: utf-8 -*-
-import pytest
-
 """Tests for bundled transliterators."""
 
-# from graphtransliterator.exceptions import IncompleteGraphCoverageException
-# from graphtransliterator.transliterators import Bundled
+import pytest
 import graphtransliterator.transliterators as transliterators
 
-# from io import StringIO
-# import pytest
-# import yaml
-import pdb
 
-pdb.set_trace()
-transliterator = transliterators.Example()  # Load example bundled transliterator
+def _test_bundled_transliterator(transliterator):
+    # Confirm transliteratorcan be loaded from yaml (descriptive) and json (faster)
+    for method in ("yaml", "json"):
 
-#
-# def test_coverage_exception():
-#     """Because all edges and nodes have been visited, there should be an exception.
-#
-#     This function can be cut from tests, but is given here as an example."""
-#     with pytest.raises(IncompleteGraphCoverageException):
-#         transliterator.graph.check_coverage()
-#
-#
-# def test_clear_coverage():
-#     """Test clearing of visit history."""
-#     transliterator.graph.clear_visited()
-#     with pytest.raises(IncompleteGraphCoverageException):
-#         transliterator.graph.check_coverage()
-#
-#     transliterator.transliterate("a ")  # run all rules
-#     assert transliterator.graph.check_coverage()
-#
-#     transliterator.graph.clear_visited()
-#     with pytest.raises(IncompleteGraphCoverageException):
-#         transliterator.graph.check_coverage()
+        def init_test_class(self, method=method, coverage=True):
+            """Initializes test class. Sets directory and name to superclass's."""
+            self._module_dir = transliterator._module_dir
+            self._module_name = transliterator._module_name
+            self.init_from(method, check_ambiguity=True, coverage=True)
 
-#
-# # Define transliteration tests
-#
-# # These should be a dictionary from source to correct target:
-# transliteration_tests = {"a": "A", " ": " "}
-#
-# # These can also come from a YAML string:
-# transliteration_tests = yaml.safe_load(
-#     StringIO(
-#         """
-#         a: A
-#         ' ': ' '
-#         """
-#     )
-# )
-# # Or use the `Bundled` class to load
-# import pdb; pdb.set_trace()
-# transliteration_tests = transliterator.load_yaml_tests()
-#
-# print(transliteration_tests)
-#
+        class_name = "Test{}From{}".format(
+            type(transliterator).__name__, method.upper()
+        )
+        SuperClass = transliterator.__class__
+        TestClass = type(class_name, (SuperClass,), {"__init__": init_test_class})
+        # Confirm it is a subclass of super class
+        assert issubclass(TestClass, SuperClass)
+        # Check no coverage works
+        assert TestClass(method="json", coverage=False)
+        # Create instance of class
+        test_class = TestClass(method=method)
+        # Load transliteration tests (in tests/test_NAME.yaml)
+        transliteration_tests = test_class.load_yaml_tests()
+        # Check that bundled tests are in valid format dict {str:str}
+        assert type(transliteration_tests) == dict
+        for k, v in transliteration_tests.items():
+            assert type(k) == str and type(v) == str
+        assert len(transliteration_tests) > 0
+        # Run yaml tests, calls run_tests()
+        test_class.run_yaml_tests()
+        # Check there is coverage of graph and onmatch rules in tests
+        assert test_class.check_coverage()
 
 
-def test_transliterator():
-    """Write tests here to confirm the transliterator works correctly.
-
-    All nodes, edges, and onmatch_rules should be visited."""
-
-    transliterator.run_yaml_tests()
+def test_bundled_transliterators():
+    for transliterator in transliterators.iter_transliterators():
+        _test_bundled_transliterator(transliterator)
 
 
-def test_duplicate_transliterator():
+def test_duplicate_transliterator_error():
     """Test that there can be no transliterators with the same name."""
     with pytest.raises(ValueError):
         transliterators.add_transliterators(path=transliterators.__path__)
@@ -80,10 +56,7 @@ def test_iter_names():
 
 
 def test_iter_transliterators():
-    """Test transliteratorss.iter_transliterators()"""
-    import pdb
-
-    pdb.set_trace()
+    """Test transliterators.iter_transliterators()"""
     example = [
         _
         for _ in transliterators.iter_transliterators()
@@ -92,52 +65,17 @@ def test_iter_transliterators():
     assert example.transliterate("a") == "A"
 
 
+def test_transliterators_metadata():
+    """Confirm that metadata of bundled transliterators matches schema."""
+    for _ in transliterators.iter_transliterators():
+        assert transliterators.MetadataSchema().load(_.metadata)
+
+
 def test_bundled():
     # check that Bundled exists
     assert transliterators.Bundled
 
 
-def test_all_transliterators():
-    import pdb
-
-    pdb.set_trace()
-    for transliterator in transliterators.iter_transliterators():
-        import pdb
-
-        pdb.set_trace()
-        test_class = type(transliterator.__class__)
-
-        x = temp_class().from_YAML(coverage=True)
-        # x = transliterator.from_YAML(coverage=True)
-        #
-        # assert transliterator.from_JSON(coverage=True)
-        # pass
-
-
-# Iterate through transliteration
-
-
-#
-#     for source, target in transliteration_tests.items():
-#         result = transliterator.transliterate(source)
-#         assert (
-#             transliterator.transliterate(source) == target
-#         ), 'Transliteration error: "{}" -> "{}"; should -> "{}"'.format(
-#             source, result, target
-#         )
-#     # Finally, test coverage — were all nodes and edges of the graph visited?
-#     # We are all onmatch_rules used?
-#
-#     transliterator.graph.check_coverage()
-#
-#
-
-#
-# def test_json_load(tmpdir):
-#     """Test JSON load."""
-#     import pdb; pdb.set_trace()
-#     assert transliterators.Example.init_from_JSON().transliterate('a') == 'A'
-#     # f = tmpdir.mkdir("sub").join("hello.txt")
-#     # f.write(transliterator.dump())
-#     # import pdb;pdb.set_trace()
-#     # assert Bundled.init_from_JSON(str(f)).transliterate("a") == "A"
+def test_no_coverage():
+    x = transliterators.Example()
+    x.from_JSON(coverage=False)